+++ src/asp/Camera/BundleAdjustCamera2.cc
@@ -1218,4 +1218,70 @@ void estimateGsdPerTriPoint(std::vector<std::string> const& images,
   return;  
 }
 
+// Find the average for the gsd for all pixels whose rays intersect at the given
+// triangulated point. This is used in jitter solving.
+void estimateGsdPerTriPoint(std::vector<std::string> const& images, 
+                            std::vector<vw::CamPtr>  const& cameras,
+                            asp::CRNJ                const& crn,
+                            std::set<int>            const& outliers,
+                            std::vector<double>      const& tri_points_vec, 
+                            // Output
+                            std::vector<double>     & gsds) {
+
+  // Sanity checks
+  if (crn.size() != images.size())
+    vw_throw(ArgumentErr() << "Expecting the same number of images and crn points.\n");
+  if (crn.size() != cameras.size())
+    vw_throw(ArgumentErr() << "Expecting the same number of images and cameras.\n");
+
+  // Image bboxes
+  std::vector<vw::BBox2> bboxes;
+  for (size_t i = 0; i < images.size(); i++) {
+    vw::DiskImageView<float> img(images[i]);
+    bboxes.push_back(bounding_box(img));
+  }
+        
+  int num_cameras = cameras.size();
+  int num_points  = tri_points_vec.size()/3;
+  
+  // Initialize all gsd to 0
+  gsds.resize(num_points, 0.0);
+  std::vector<int> count(num_points, 0);
+  
+  for (int icam = 0; icam < num_cameras; icam++) { // Camera loop
+    for (auto fiter = crn[icam].begin(); fiter != crn[icam].end(); fiter++) { // IP loop
+
+      // The index of the 3D point this IP is for.
+      int ipt = (**fiter).m_point_id;
+      
+      VW_ASSERT(int(icam) < num_cameras,
+                ArgumentErr() << "Out of bounds in the number of cameras.");
+      VW_ASSERT(int(ipt)  < num_points,
+                ArgumentErr() << "Out of bounds in the number of points.");
+      
+      if (outliers.find(ipt) != outliers.end())
+        continue; // Skip outliers
+
+      vw::Vector2 pix = (**fiter).m_location;
+      
+      double const* point = &tri_points_vec[3*ipt];
+      Vector3 xyz(point[0], point[1], point[2]);
+
+      // Estimate the GSD at the given pixel given an estimate of the ground point
+      double gsd = vw::camera::estimatedGSD(cameras[icam].get(), bboxes[icam], 
+                                            pix, xyz);
+      gsds[ipt] += gsd;
+      count[ipt]++;
+    }
+  }
+  
+  // Find the average gsd
+  for (int ipt = 0; ipt < num_points; ipt++) {
+    if (count[ipt] > 0)
+      gsds[ipt] /= count[ipt];
+  }
+  
+  return;  
+}
+
 } // end namespace asp